test(e2e): isolate env via .env.test and mock S3 with MinIO
기존 `pnpm test:e2e`는 `cp .env.example .env`가 무조건 실행돼 개발자의
로컬 `.env`를 placeholder(`postgres://user:password@host:5432/schema`)로
덮어썼고, 그 결과 migrate/seed가 존재하지 않는 DB로 붙으려 실패하고
있었습니다. 이 커밋은 e2e 파이프라인 전체를 재정비합니다.

- `.env.test` 도입 (dummy 값만 담겨있어 커밋). Playwright, prepare,
  seed 스크립트 모두 이 파일만 명시적으로 로드하므로 개발자 로컬
  `.env`를 건드리지 않습니다. `.gitignore`에 `!.env.test` 예외 추가.
- `playwright.config.ts`의 webServer `env:` 블록 제거 — Playwright가
  자동으로 `process.env`를 상속하므로 중복.
- `package.json`의 `test:e2e`, `test:e2e:ui`에서 `cp .env.example .env`
  제거.
- `docker-compose.test.yml`에 MinIO 서비스 추가. entrypoint를 override해
  `mkdir -p /data/stdev-kr && exec minio server /data`로 버킷 디렉토리를
  먼저 만든 뒤 서버를 기동 — 별도 init 사이드카 없이 한 서비스로 해결.
- `src/utils/s3.ts`: AWS SDK 공식 env var인 `AWS_ENDPOINT_URL_S3`가
  세팅되면 `endpoint + forcePathStyle: true`를 `S3Client`에 주입.
  프로덕션에는 이 변수를 세팅하지 않으므로 동작 변화 없음. `.env.test`에
  `AWS_ENDPOINT_URL_S3=http://127.0.0.1:9000` 포함.
- `scripts/` → `src/scripts/`로 이동 (git rename으로 이력 보존).
  `src/utils/`(런타임 헬퍼)와 혼동되지 않도록 `src/tests/`, `src/e2e/`
  와 평행한 형제 디렉토리로 배치.
- `package.json`의 stale `migrate:payload` 엔트리 제거 — 참조하던 파일
  `scripts/migrate-payload-to-prisma.ts`는 이미 커밋 534f216에서
  PayloadCMS 정리와 함께 삭제됨.
- `codecov.yml`의 ignore 경로를 `src/scripts/**/*`로 갱신.
- `src/scripts/prepare-playwright-e2e.mjs`: 커스텀 pg_isready 폴링 루프
  제거, `docker compose up -d --wait`로 postgres와 minio 둘 다
  healthcheck 기반 대기.
- README의 E2E 섹션에 MinIO 의존성 문구 추가.
