+++ .env.test
@@ -1,0 +1,24 @@
+# E2E test env loaded by Playwright, prepare script, and seed script.
+# Values must match docker-compose.test.yml (postgres + minio services).
+# Safe to commit — everything here is a dummy credential.
+
+DATABASE_URL=postgresql://stdev:stdev@127.0.0.1:5433/stdev_test
+
+BETTER_AUTH_SECRET=test-better-auth-secret-32-chars-min-abcd
+BETTER_AUTH_URL=http://127.0.0.1:3100
+
+GOOGLE_CLIENT_ID=test-google-client-id
+GOOGLE_CLIENT_SECRET=test-google-client-secret
+
+NEXT_PUBLIC_CHANNEL_PLUGIN_KEY=test-channel-key
+NEXT_PUBLIC_GTM_ID=GTM-TEST
+NEXT_PUBLIC_GA_ID=G-TEST
+
+AWS_REGION=ap-northeast-2
+AWS_ACCESS_KEY=test-access-key
+AWS_SECRET_KEY=test-secret-key
+S3_BUCKET=stdev-kr
+
+# Point @aws-sdk/client-s3 at the MinIO sidecar from docker-compose.test.yml.
+# In production this var is unset, so S3Client uses the real AWS endpoint.
+AWS_ENDPOINT_URL_S3=http://127.0.0.1:9000
