+++ .gitignore
@@ -37,6 +37,7 @@ yarn-error.log*
 # env files (can opt-in for commiting if needed)
 .env*
 !.env.example
+!.env.test
 
 # vercel
 .vercel
